Исправление ошибок в случае невозможности загрузки точек маршрута, офферов или информации о пунктах назначения
