+++ src/utils/const.js
@@ -1,14 +1,11 @@
-import { DateFormats } from './utils.js';
-
 const END_POINT = 'https://21.objects.pages.academy/big-trip';
 const AUTH_TOKEN = 'Basic 44b7-b4ca-e34f0eb25871';
 const RANDOM_PHOTOS_SERVICE_URL = 'https://loremflickr.com/248/152?random=';
-const CITY_NAMES = ['Moskow', 'London', 'Amsterdam', 'New Zealand', 'Switzerland', 'China', 'Japan'];
-const OFFER_NAMES = ['Transfer', 'Meet in Airport', 'Extra Luggage', 'Lunch', 'Switch to comfort'];
 const POINT_TYPES = ['Taxi', 'Bus', 'Train', 'Ship', 'Drive', 'Flight', 'Check-in', 'Sightseeing', 'Restaurant'];
+const DEFAULT_POINT_TYPE = 'Flight';
 const BLANK_POINT = {
   id: crypto.randomUUID(),
-  type: POINT_TYPES[5],
+  type: DEFAULT_POINT_TYPE,
   destination: '',
   dates: {
     start: '',
@@ -18,7 +15,19 @@ const BLANK_POINT = {
   cost: 0,
   isFavorite: false,
 };
-const POINTS_COUNT = 4;
+const DateFormats = {
+  FLATPICKR: 'd/m/y H:i', // Для Флэтпикера
+  DATE_TIME: 'YYYY-MM-DD[T]hh:mm', // Для тега datetime
+  PATH: 'DD/MM/YY',
+  DAY: 'DD',
+  CHOSED_DATE: 'DD/MM/YY HH:mm', // Дата и время начала события
+  FOR_POINT_PERIODS: 'HH:mm', // Для периодов, выбранных для точки маршрута
+  FOR_POINT: 'MMM DD', // Дата для каждой конкретной точки маршрута
+  // Форматирование продолжительности нахождения в точке маршрута
+  LESS_THAN_HOUR: 'mm', // Менее часа
+  LESS_THAN_DAY: 'HH mm', // Менее суток
+  MORE_THAN_DAY: 'DD HH mm' // Более суток
+};
 const FLATPIKR_SETTINGS = {
   enableTime: true,
   dateFormat: DateFormats.FLATPICKR,
@@ -62,18 +71,16 @@ const TRIP_INFO_TEMPLATE = '<section class="trip-main__trip-info  trip-info"></s
 export {
   END_POINT,
   AUTH_TOKEN,
-  CITY_NAMES,
   RANDOM_PHOTOS_SERVICE_URL,
-  OFFER_NAMES,
   POINT_TYPES,
   BLANK_POINT,
-  POINTS_COUNT,
   FLATPIKR_SETTINGS,
   API_URL,
-  Method,
   TRIP_EVENTS_LIST_TEMPLATE,
   TRIP_INFO_TEMPLATE,
+  DateFormats,
   ActionType,
   UpdateType,
-  TripBoardMode
+  TripBoardMode,
+  Method,
 };
